Extract shared methods into a BaseStrategy class
