+++ lib/brainstem.rb
@@ -2,6 +2,9 @@ require "brainstem/version"
 require "brainstem/presenter"
 require "brainstem/presenter_collection"
 require "brainstem/controller_methods"
+require "brainstem/query_strategies/base_strategy"
+require "brainstem/query_strategies/filter_and_search"
+require "brainstem/query_strategies/filter_or_search"
 
 # The Brainstem module itself contains a +default_namespace+ class attribute and a few helpers that make managing +PresenterCollections+ and their corresponding namespaces easier.
 module Brainstem
