+++ lib/brainstem/query_strategies/filter_or_search.rb
@@ -1,10 +1,6 @@
 module Brainstem
   module QueryStrategies
-    class FilterOrSearch
-      def initialize(options)
-        @options = options
-      end
-
+    class FilterOrSearch < BaseStrategy
       def execute(scope)
         if searching?
           # Search
@@ -81,24 +77,6 @@ module Brainstem
         end
       end
 
-      def calculate_limit
-        [[@options[:params][:limit].to_i, 1].max, (@options[:max_per_page] || @options[:default_max_per_page]).to_i].min
-      end
-
-      def calculate_offset
-        [@options[:params][:offset].to_i, 0].max
-      end
-
-      def calculate_per_page
-        per_page = [(@options[:params][:per_page] || @options[:per_page] || @options[:default_per_page]).to_i, (@options[:max_per_page] || @options[:default_max_per_page]).to_i].min
-        per_page = @options[:default_per_page] if per_page < 1
-        per_page
-      end
-
-      def calculate_page
-        [(@options[:params][:page] || 1).to_i, 1].max
-      end
-
       def paginate(scope)
         if @options[:params][:limit].present? && @options[:params][:offset].present?
           limit = calculate_limit
@@ -129,18 +107,6 @@ module Brainstem
         end
 
         ordered_records.compact
-      end
-
-      def filter_includes
-        allowed_associations = @options[:primary_presenter].allowed_associations(@options[:params][:only].present?)
-
-        [].tap do |selected_associations|
-          (@options[:params][:include] || '').split(',').each do |k|
-            if association = allowed_associations[k]
-              selected_associations << association
-            end
-          end
-        end
       end
     end
   end
